Section 25: Working with REST API (frontend and backed)

+++ Section 25 - Working with REST API - BackEnd/controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 
 const { validationResult } = require('express-validator/check');
 const Post = require('../models/post');
+const User = require('../models/user');
 //* http method and the name
 //* Adding pagination
 exports.getPosts = (req, res, next) => {
@@ -72,22 +73,29 @@ exports.createPost = (req, res, next) => {
     //* Create a post
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     //* Create a new post using mongoDb and the model
     const post = new Post({ 
         title: title, 
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'David' 
-        }
+        creator: req.userId
     });
     //* method save provided by mongoose to save data into the data base
     post.save()
     .then(result => {
-        console.log(result);
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post created successfully',
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name }
         });
     })
     .catch(err => {
@@ -141,8 +149,13 @@ exports.updatePost = (req, res, next) => {
     Post.findById(postId)
     .then(post => {
         if(!post) {
-            const error = new Error('Could not finf post.');
+            const error = new Error('Could not find post.');
             error.statusCode = 404;
+            throw error;
+        }
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('No Authorized!');
+            error.statusCode = 403;
             throw error;
         }
         if (imageUrl !== post.imageUrl) {
@@ -168,26 +181,37 @@ exports.deletePost = (req, res, next) => {
     const postId = req.params.postId;
     Post.findById(postId)
         .then(post => {
-            //* Check logged in user
-            if(!post) {
-                const error = new Error('Could not finf post.');
+            if (!post) {
+                const error = new Error('Could not find post.');
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
+        // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Deleted post.' })
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
+            res.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
-            if(!err.statusCode) {
+            if (!err.statusCode) {
                 err.statusCode = 500;
             }
             next(err);
-        })
-}
+        });
+};
 
 const clearImage = filePath => {
     filePath = path.join(__dirname, '..', filePath);
